requirements.txt
aiogram==3.0.0b7
alembic==1.10.2
asyncpg==0.27.0
sqlalchemy[asyncio]==2.0.15
python-dotenv==1.0.0
psycopg2-binary==2.9.6